+++ src/sisl/io/openmx/md.py
@@ -15,9 +15,7 @@ __all__ = ["mdSileOpenMX"]
 
 @set_module("sisl.io.openmx")
 class mdSileOpenMX(xyzSile, SileOpenMX):
-
-    def read_geometry(self, *args, all=True, **kwargs):
-        return super().read_geometry(*args, all=all, **kwargs)
+    pass
 
 
 add_sile('md', mdSileOpenMX, gzip=True)