+++ src/sisl/io/siesta/ani.py
@@ -15,9 +15,7 @@ __all__ = ["aniSileSiesta"]
 
 @set_module("sisl.io.siesta")
 class aniSileSiesta(xyzSile, SileSiesta):
-
-    def read_geometry(self, *args, all=True, **kwargs):
-        return super().read_geometry(*args, all=all, **kwargs)
+    pass
 
 
 add_sile('ANI', aniSileSiesta, gzip=True)