+++ src/sisl/io/tests/test_xsf.py
@@ -57,7 +57,7 @@ def test_axsf_geoms(sisl_tmp):
             s.write_geometry(g)
 
     with xsfSile(f) as s:
-        rgeoms = s.read_geometry(start=0, stop=2)
+        rgeoms = s.read_geometry[0:2]()
         assert len(rgeoms) == 2
         assert all(isinstance(rg, Geometry) for rg in rgeoms)
         print()
@@ -67,18 +67,18 @@ def test_axsf_geoms(sisl_tmp):
         assert all(g.equal(rg) for g, rg in zip(geoms, rgeoms))
 
     with xsfSile(f) as s:
-        rgeoms = s.read_geometry(start=1)
+        rgeoms = s.read_geometry[1]()
         assert isinstance(rgeoms, Geometry)
         assert geoms[1].equal(rgeoms)
 
     with xsfSile(f) as s:
-        rgeoms = s.read_geometry(all=True)
+        rgeoms = s.read_geometry[:]()
         assert len(rgeoms) == len(geoms)
         assert all(isinstance(rg, Geometry) for rg in rgeoms)
         assert all(g.equal(rg) for g, rg in zip(geoms, rgeoms))
 
     with xsfSile(f) as s:
-        rgeoms, rdata = s.read_geometry(all=True, ret_data=True)
+        rgeoms, rdata = s.read_geometry[:](ret_data=True)
         assert len(rgeoms) == 3
         assert all(g.equal(rg) for g, rg in zip(geoms, rgeoms))
         for dat in rdata:
@@ -96,13 +96,13 @@ def test_axsf_data(sisl_tmp):
             s.write_geometry(g, data=dat)
 
     with xsfSile(f) as s:
-        rgeoms, rdata = s.read_geometry(all=True, ret_data=True)
+        rgeoms, rdata = s.read_geometry[:](ret_data=True)
         assert len(rgeoms) == len(geoms)
         assert all(g.equal(rg) for g, rg in zip(geoms, rgeoms))
         assert all(np.allclose(d0, d1) for d0, d1 in zip(rdata, data))
 
     with xsfSile(f) as s:
-        rgeoms, rdata = s.read_geometry(start=0, ret_data=True)
+        rgeoms, rdata = s.read_geometry[0](ret_data=True)
         assert geoms[0].equal(rgeoms)
         assert np.allclose(rdata, data[0])